chore: Refactor import statements and function names in pdfdeal module
