+++ src/pdfdeal/__init__.py
@@ -1,3 +1,3 @@
-from .get_file import deal_pdf
+from .dealpdfs import deal_pdf
 
 __all__ = ['deal_pdf']
