+++ src/pdfdeal/doc2x.py
@@ -2,7 +2,7 @@ import asyncio
 import os
 from .Doc2X.Exception import RateLimit, RateLimiter
 from .Doc2X.Types import OutputFormat, OutputVersion, RAG_OutputType
-from .get_file import strore_pdf
+from .dealpdfs import strore_pdf
 from typing import Tuple
 from .file_tools import list_rename
